+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ## Schedule
 
+### Release `0.0.2`
+
+1. Removes 'prefix' configuration and adds Key Mapping.
+1. Integrates with Travis CI.
+
 ### Release `0.0.1`
 
-  * Exports function `rat_error` (after calling `use RatError`).
-  * Exports `rat_error` configuration (see `config/*.exs` for detail).
-  * Implements module `RatError.Formatter` and `RatError.Structure`.
+1. Exports function `rat_error` (after calling `use RatError`).
+1. Exports `rat_error` configuration (see `config/*.exs` for detail).
+1. Implements module `RatError.Formatter` and `RatError.Structure`.